service/ingestion-client/requirements.txt
flask
requests
grpcio
grpcio-tools
